vride/src/main/java/com/carpooling/vride/model/Rider.java
package com.carpooling.vride.model;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.Table;
@Entity
@Table(name="riders")
public class Rider {
	@Id
	@Column(name="empid")
	private int empid;
    @Column(name="username")
	private String username;
    @Column(name="origin")
    private String origin;
    @Column(name="destination")
    private String destination;

	public int getEmpid() {
			return empid;
		}
		public void setEmpid(int empid) {
			this.empid = empid;
		}
		public String getUsername() {
			return username;
		}
		public void setUsername(String username) {
			this.username = username;
		}
		public String getOrigin() {
			return origin;
		}
		public void setOrigin(String origin) {
			this.origin = origin;
		}
		public String getDestination() {
			return destination;
		}
		public void setDestination(String destination) {
			this.destination = destination;
		}
		
}
